requirements.txt
sly>=0.4
lotus-ai
